+++ src/ZLinq/Linq/Shuffle.cs
@@ -73,7 +73,7 @@ namespace ZLinq.Linq
         {
             InitBuffer();
 
-            if (index < buffer.Length)
+            if ((uint)index < (uint)buffer.Length)
             {
                 current = buffer.UnsafeGetAt(index);
                 index++;
